overlap fused marlin moe silu_and_mul and zeros
